+++ src/client/java/rh/maparthelper/conversion/CroppingMode.java
@@ -3,5 +3,6 @@ package rh.maparthelper.conversion;
 public enum CroppingMode {
     NO_CROP,
     AUTO_CROP,
+    FIT,
     USER_CROP
 }